Only remove sources if they exist

+++ tools/src/sources.rs
@@ -88,7 +88,9 @@ async fn extract_source(name: &str, source: &Source) -> Result<()> {
         return Ok(());
     }
 
-    fs::remove_dir_all(archive_path)?;
+    if archive_path.exists() {
+        fs::remove_dir_all(archive_path)?;
+    }
 
     let target = File::open(target_path)?;
